add shared template registry, update publish form, add email template info
